requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.3
django-debug-toolbar==4.0.0
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mutagen==1.46.0
oauthlib==3.2.2
packaging==21.3
phonenumberslite==8.13.5
Pillow==9.3.0
psycopg==3.1.4
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13

celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1